Moved markdownlint loadConfig function to a dedicated file 📁
